Write coordinate data in one go when possible
When we use the Adios2StMan to bind all columns of the MeasurementSet we
gain the ability to write the coordinate data in a single writing
operation instead of doing it in a per-row basis. To get the full
performance one also needs the latest Adios2StMan that implements the
scalar column-wise read/write operations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>